removed stop
According to the data sheet,  a stop is not needed between the write/read when reading a word from the RAM or EEPROM.  This caused very subtle issues which would occasionally cause the sensor to report 0xFF 0xFF as the temperature.   It may also be the reason SF developers saw strange behavior when changing the bitrate.